fix: rename component and imports
